+++ src/operations.ts
@@ -1,4 +1,0 @@
-
-const Operations = [ "select", "insert", "update", "merge", "delete" ];
-
-export { Operations };