+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # 0.6.0 (Mar 29th 2025)
 
 - Add getters for table, unsafeProps, cols, properties
-- Delete getTabe(), getCols() and getUnsafeProps() methods
-- Add getColsByOp() method to get cols for a specific operation
 - Add getProp() method to retreive a property with its key
+- Delete getTabe() and getUnsafeProps() methods
+- Delete cols parameter. Now used in pgsql plugin
+- Replace operations by REST methods
 
 # 0.5.0 (Jan 11th 2025)